bai2

+++ JAVA/lesson1/Hello World/src/App.java
@@ -3,11 +3,11 @@ import java.util.Scanner;
 public class App {public static void main(String[] args) throws Exception {
         Scanner sc = new Scanner(System.in);
         System.out.println("Nhập cân nặng theo kg: ");
-        float weight = sc.nextFloat();
+        double weight = sc.nextFloat();
         System.out.println("Nhập chiều cao theo m: ");
-        float height = sc.nextFloat();
+        double height = sc.nextFloat();
 
-        // float bmi = weight / (height * height);
+        // double bmi = weight / (height * height);
         // System.out.println("BMI là: "+ bmi);
         System.out.println("BMI là: " + weight / (height * height));
         sc.close();
